requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
requests>=2.31.0
psutil>=5.9.0
aiofiles>=23.0.0
piexif>=1.1.3
